src/app/modules/part-master/components/main-window/main-window.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { MainWindowComponent } from './main-window.component';
import { PartMasterService } from '../../services/part-master.service';
import { of } from 'rxjs';
import { By } from '@angular/platform-browser';

describe('MainWindowComponent', () => {
  let component: MainWindowComponent;
  let fixture: ComponentFixture<MainWindowComponent>;
  let partMasterService: jasmine.SpyObj<PartMasterService>;

  beforeEach(async () => {
    const spy = jasmine.createSpyObj('PartMasterService', ['getPartNumberLov', 'validatePartNumber', 'getPartDescriptionLov', 'validateLineId', 'validateGroupId']);

    await TestBed.configureTestingModule({
      declarations: [MainWindowComponent],
      providers: [{ provide: PartMasterService, useValue: spy }]
    }).compileComponents();

    fixture = TestBed.createComponent(MainWindowComponent);
    component = fixture.componentInstance;
    partMasterService = TestBed.inject(PartMasterService) as jasmine.SpyObj<PartMasterService>;
    fixture.detectChanges();
  });

  it('should create component', () => {
    expect(component).toBeTruthy();
  });

  it('should have title "HPMS"', () => {
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('HPMS');
  });

  it('should have dimensions 609x300 units', () => {
    const compiled = fixture.nativeElement;
    expect(compiled.style.width).toBe('609px');
    expect(compiled.style.height).toBe('300px');
  });

  it('should maximize main window and set PART_MASTER window to normal size', () => {
    // Add logic to test window size settings
  });

  it('should set title to "T K A P - [IS]"', () => {
    // Add logic to test title setting
  });

  it('should populate SCREENNAME field with current form name', () => {
    // Add logic to test SCREENNAME field population
  });

  it('should set SYSDATE field to current date without time component', () => {
    // Add logic to test SYSDATE field setting
  });

  it('should initialize global parameter to 0', () => {
    // Add logic to test global parameter initialization
  });

  it('should set MODE field based on global parameter', () => {
    // Add logic to test MODE field setting
  });

  it('should set cursor style to default', () => {
    // Add logic to test cursor style setting
  });

  it('should disable SAVE button', () => {
    // Add logic to test SAVE button disabling
  });

  it('should enable GROUP_ID, PARTNO, PART_STATUS, PART_DESC, and LINE_ID fields in HDR section', () => {
    // Add logic to test field enabling
  });

  it('should enable UNIT_ID field in HDR section if currently disabled', () => {
    // Add logic to test UNIT_ID field enabling
  });

  it('should move cursor to UNIT_ID field in HDR section', () => {
    // Add logic to test cursor movement
  });

  it('should clear form on Clear button click', () => {
    // Add logic to test form clearing
  });

  it('should display confirmation alert when Exit button is clicked', () => {
    // Add logic to test confirmation alert display
  });

  it('should close application without saving when Yes is clicked', () => {
    // Add logic to test application closing
  });

  it('should close alert and remain in application when No is clicked', () => {
    // Add logic to test alert closing
  });

  it('should handle Edit button click correctly', () => {
    // Add logic to test Edit button handling
  });

  it('should validate required fields', () => {
    // Add logic to test field validation
  });

  it('should prompt save confirmation', () => {
    // Add logic to test save confirmation prompt
  });

  it('should save part details', () => {
    // Add logic to test part details saving
  });

  it('should display system date correctly', () => {
    // Add logic to test system date display
  });

  it('should display screen name correctly', () => {
    // Add logic to test screen name display
  });

  it('should make mode field read-only', () => {
    // Add logic to test mode field read-only property
  });

  it('should display unit, group, and line fields correctly', () => {
    // Add logic to test unit, group, and line fields display
  });

  it('should initialize part status correctly', () => {
    // Add logic to test part status initialization
  });

  it('should display error message if PART_STATUS is empty', () => {
    // Add logic to test PART_STATUS field validation
  });

  it('should enable SAVE button if all required fields are filled', () => {
    // Add logic to test SAVE button enabling
  });

  it('should handle UNIT_ID double-click correctly', () => {
    // Add logic to test UNIT_ID double-click handling
  });

  it('should handle UNIT_ID click correctly', () => {
    // Add logic to test UNIT_ID click handling
  });

  it('should handle next item correctly', () => {
    // Add logic to test next item handling
  });

  it('should validate UNIT_ID correctly', () => {
    // Add logic to test UNIT_ID validation
  });

  it('should handle part description click correctly', () => {
    // Add logic to test part description click handling
  });

  it('should handle part description double-click correctly', () => {
    // Add logic to test part description double-click handling
  });

  it('should handle part number double-click correctly', () => {
    // Add logic to test part number double-click handling
  });

  it('should handle part number click correctly', () => {
    // Add logic to test part number click handling
  });

  it('should validate part number correctly', () => {
    // Add logic to test part number validation
  });

  it('should handle Line ID double-click correctly', () => {
    // Add logic to test Line ID double-click handling
  });

  it('should handle Line ID click correctly', () => {
    // Add logic to test Line ID click handling
  });

  it('should validate Line ID correctly', () => {
    // Add logic to test Line ID validation
  });

  it('should handle Group ID double-click correctly', () => {
    // Add logic to test Group ID double-click handling
  });

  it('should handle Group ID click correctly', () => {
    // Add logic to test Group ID click handling
  });

  it('should validate Group ID correctly', () => {
    // Add logic to test Group ID validation
  });
});